+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.10.0
+### Changed
+- Move modules out of Network.Haskoin namespace.
+- Add better and more logging.
+- Change Manager module name and related values to PeerManager.
+
 ## 0.9.21
 ### Removed
-- Removed unnecessary logging.
+- Remove unnecessary logging.
 
 ## 0.9.20
 ### Changed